requirements.txt
# TTS Server Dependencies
flask>=3.0.0
chatterbox-tts
torch
torchaudio
